moved category-navbar to shared module and added structural directive to dynamically render the navbar depending on route
