--- app/shared/components/footer/footer.component.html
@@ -1,18 +1,0 @@
-<footer class="sticky-bottom">
-
-    <div class="container">
-        <div class="col-md-12 text-center">
-
-          <div class="logo">NG-Blog</div>
-            <ul class="nav justify-content-center">
-              <li class="nav-item"><a href="" class="nav-link">Home</a></li>
-              <li class="nav-item"><a href="" class="nav-link">About</a></li>
-              <li class="nav-item"><a href="" class="nav-link">Terms & Conditions</a></li>
-              <li class="nav-item"><a href="" class="nav-link">Contact</a></li>
-            </ul>
-            <p class="mb-2 mt-2">Angular Blogsite 2024</p>
-
-        </div>
-      </div>
-
-  </footer>
