--- app/shared/components/footer/footer.component.scss
@@ -1,13 +1,0 @@
-@import "../../../../styles";
-footer {
-  height: 130px;
-  padding: 15px;
-  background-color: $navbar-footer-color;
-}
-
-.logo {
-  font-size: 1.2em;
-}
-.nav-link {
-  color: $primary-color;
-}
