--- app/shared/components/header/header.component.html
@@ -1,6 +1,0 @@
-<header class="navbar navbar-expand-lg bg-white">
-  <div class="container justify-content-center">
-    <a href="#" class="logo">NG-Blog</a>
-  </div>
-</header>
-
