--- app/shared/components/header/header.component.scss
@@ -1,9 +1,0 @@
-@import '../../../../styles';
-
-// header {
-//   height: 100px;
-// }
-
-.logo {
-  font-size: 1.5em;
-}
