+++ app/shared/shared.module.ts
@@ -1,8 +1,9 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { FooterComponent } from './components/footer/footer.component';
-import { HeaderComponent } from './components/header/header.component';
+import { FooterComponent } from './components/layout/footer/footer.component';
+import { HeaderComponent } from './components/layout/header/header.component';
+import { CategoryNavbarComponent } from './components/layout/category-navbar/category-navbar.component';
 
 
 
@@ -10,6 +11,7 @@ import { HeaderComponent } from './components/header/header.component';
   declarations: [
     HeaderComponent,
     FooterComponent,
+    CategoryNavbarComponent,
   ],
   imports: [
     CommonModule
@@ -17,6 +19,7 @@ import { HeaderComponent } from './components/header/header.component';
   exports:[
     HeaderComponent,
     FooterComponent,
+    CategoryNavbarComponent,
   ],
 })
 export class SharedModule { }
