--- app/user/components/layout/category-navbar/category-navbar.component.html
@@ -1,12 +1,0 @@
-<nav class="navbar navbar-expand-lg ">
-  <div class="container-fluid justify-content-center">
-      <ul class="navbar-nav">
-        <li class="nav-item ms-3"><a href="" class="nav-link">Desgin</a></li>
-        <li class="nav-item ms-3"><a href="" class="nav-link">Inspiration</a></li>
-        <li class="nav-item ms-3"><a href="" class="nav-link">Sports</a></li>
-        <li class="nav-item ms-3"><a href="" class="nav-link">Politics</a></li>
-      </ul>
-    </div>
-  </nav>
-
-  <!-- <div class="collapse navbar-collapse "></div> -->
