--- app/user/components/layout/category-navbar/category-navbar.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'app-category-navbar',
-  templateUrl: './category-navbar.component.html',
-  styleUrls: ['./category-navbar.component.scss']
-})
-export class CategoryNavbarComponent {
-
-}
